Rename `coords` module to `coord` across the codebase.
This refactor standardizes the naming of the `lib.geo.valueobject` module by renaming `coords.py` to `coord.py`. All corresponding imports were updated to maintain consistency and avoid import errors. No functional changes were introduced.
