+++ soil_analysis/domain/valueobject/land.py
@@ -1,4 +1,4 @@
-from lib.geo.valueobject.coords import LandCoords
+from lib.geo.valueobject.coord import LandCoords
 
 
 class Land:
